[RFR] fetch..folder: put each test class in own file
