+++ fetchmail_from_imap_folder/__init__.py
@@ -1,5 +1,5 @@
-# Copyright - 2013-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 from . import match_algorithm
 from . import models
 from . import wizard
