+++ fetchmail_from_imap_folder/models/__init__.py
@@ -1,4 +1,4 @@
-# Copyright - 2013-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 from . import fetchmail_server
 from . import fetchmail_server_folder
