+++ fetchmail_from_imap_folder/tests/__init__.py
@@ -1,3 +1,4 @@
-# Copyright - 2015-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import test_attach_mail_manually
 from . import test_match_algorithms
