+++ fetchmail_from_imap_folder/wizard/__init__.py
@@ -1,3 +1,3 @@
-# Copyright - 2013-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 from . import attach_mail_manually
